+++ client/src/services/injections.ts
@@ -1,11 +1,10 @@
 import type { GeneralSocketC } from '@/services/socket';
-import type { gamesSummary, subscriptionState } from './useGameView';
+import type { gamesSummary } from './useGamesSummary';
 import { InjectionKey, inject } from 'vue';
 import { friendsState } from './useFriends';
 
 export const SocketKey: InjectionKey<GeneralSocketC> = Symbol('socket');
 export const GamesSummaryKey: InjectionKey<gamesSummary> = Symbol('gamesSummary');
-export const SubscriptionStateKey: InjectionKey<subscriptionState> = Symbol('subscriptionState');
 export const FriendsStateKey: InjectionKey<friendsState> = Symbol('friendsState');
 
 export function injectStrict<T>(key: InjectionKey<T>, fallback?: T): T {
